requirements.txt
openai
requests
streamlit
python-dotenv